Fetch CRAN datasets in separate folder.
Common fetch functions are in a separate module again.
